dom/base/test/jsmodules/test_moduleNotFound.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test what happens when a top-level module is not found</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var wasRun = false;
  var errorCount = 0;
  var eventCount = 0;

  SimpleTest.waitForExplicitFinish();

  window.onerror = function(message, url, line, column, error) {
    errorCount++;
  }

  function testError() {
    ok(eventCount == 1, 'Check that an error event was fired');
    ok(errorCount == 0, 'Check that no error was reported');
    SimpleTest.finish();
  }
</script>
<script type="module" src="./nonExistentModule.js" onerror="eventCount++">
</script>
<body onload='testError()'></body>
